chore: move files to specialized folders and remove redundant profileId check
