+++ integrations/klaviyo/definitions/schemas/regex.ts
